tests: moved all admin related commands to its own module
